refactor(coverage): move re-usable parts to base provider
